# src/main/java/com/haroun/book_network/feedback/FeedbackResponse.java
package com.haroun.book_network.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}